Use dracut by default in SUSE 550_rebuild_initramfs.sh (#3155)
Overhauled finalize/SUSE_LINUX/i386/550_rebuild_initramfs.sh
Now it uses dracut by default and mkinitrd as fallback
see https://github.com/rear/rear/issues/3152
Additionally improved the user messages (in particular the warnings)
to make it more clear that the point is to decide if the recreated system
will boot with the initrd 'as is' from the backup restore,
cf. https://github.com/rear/rear/issues/3162
Furthermore removed the whole INITRD_MODULES code
because INITRD_MODULES is not used and
/etc/sysconfig/kernel does no longer exist since SLES12
so the INITRD_MODULES code is dead code.