Search for `rhel` when detecting RHEL
The `redhat` substring is present only in auxiliary variables and URLs in
`/etc/os-release` both on Fedora and RHEL.  Therefore, the `grep` worked only
by accident on RHEL and was broken on Fedora because it wrongly classified
Fedora as RHEL because its BUG_REPORT_URL refers to Red Hat's Bugzilla.

On the other hand, `rhel` is the actual `ID` of RHEL and this string is
not present in Fedora's `/etc/os-release` at all.

Resolves: https://github.com/rear/rear/issues/3149#issuecomment-1964290116